+++ Ladybug/App/TabBarController.swift
@@ -35,6 +35,18 @@ class TabBarController: UITabBarController {
 }
 
 extension TabBarController: RadarURLOpenerUI {
+    func openRadarInDetailViewController(_ radar: Radar) {
+        let vc = DetailViewController(radar: radar)
+
+        if let navController = self.selectedViewController as? UINavigationController {
+            if let vc = navController.topViewController as? DetailViewController {
+                vc.radar = radar
+            } else {
+                navController.pushViewController(vc, animated: true)
+            }
+        }
+    }
+
     func openRadarLinkInSafariViewController(_ radarNumber: RadarNumber, radarOption: RadarOption, readerMode: Bool) {
         let url = radarNumber.url(by: radarOption)
         let sfvc = safariViewController(url: url, readerMode: readerMode)