+++ Ladybug/Bookmarks/BookmarksViewController.swift
@@ -68,7 +68,7 @@ class BookmarksViewController: UITableViewController, TableViewControllerUsingVi
                     OpenRadarAPI().fetchRadar(by: radar.number) { (result) in
                         switch result {
                         case .value(let radar):
-                            RadarCollection.shared.upsert(radar: radar)
+                            _ = RadarCollection.shared.upsert(radar: radar)
                             try? RadarCollection.shared.updatedViewed(radarNumber: radar.number)
                         case .error(let error):
                             print(error.localizedDescription)