+++ Ladybug/Model/BrowserOption.swift
@@ -13,4 +13,5 @@ enum BrowserOption: Int {
     case sfvc
     case safari
     case briskApp
+    case native
 }