+++ Ladybug/ViewModels/TableViewCellViewModel.swift
@@ -47,6 +47,7 @@ struct TableViewCellViewModel {
 
     func configure(_ cell: UITableViewCell) {
         cell.textLabel?.text = title
+        cell.textLabel?.numberOfLines = 0
         cell.detailTextLabel?.text = subtitle
         cell.selectionStyle = selectionStyle
         cell.accessoryType = accessoryType